requirements.txt
# Core dependencies
flask>=2.3.0
requests>=2.31.0
geopy>=2.3.0

# Machine Learning dependencies
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Visualization (for training)
matplotlib>=3.7.0
seaborn>=0.12.0

# Twilio for SMS and IVR (optional)
twilio>=8.0.0

# Email support (built-in with Python)
# smtplib, email.mime are part of Python standard library

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Additional utilities
python-dotenv>=1.0.0  # For environment variable management
